new libft added

+++ libft/srcs/arrow_functions/ft_memrealloc_array.c
@@ -6,22 +6,22 @@
 /*   By: sschmele <sschmele@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/25 16:12:43 by sschmele          #+#    #+#             */
-/*   Updated: 2020/07/25 16:14:53 by sschmele         ###   ########.fr       */
+/*   Updated: 2021/08/09 21:26:32 by sschmele         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void				**ft_memrealloc_array(void ***subj, int len_subj,
-						int len_needed)
+void	**ft_memrealloc_array(void ***subj, size_t len_subj,
+						size_t len_needed)
 {
-	int				i;
+	size_t			i;
 	void			**new;
 	void			**old;
 
 	i = 0;
 	old = *subj;
-	new = (void**)ft_xmalloc(sizeof(void*) * (len_needed + 1));
+	new = (void **)ft_xmalloc(sizeof(void *) * (len_needed + 1));
 	while (old[i] && i < len_subj)
 	{
 		new[i] = old[i];
